+++ Common/src/test/java/com/almostreliable/unified/util/JsonCompareTest.java
@@ -8,78 +8,70 @@ import org.junit.jupiter.api.Test;
 import java.util.Arrays;
 import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 public class JsonCompareTest {
-    public static String recipe = """
-            {
-              "type": "minecraft:smelting",
-              "group": "coal",
-              "ingredient": {
-                "item": "minecraft:coal_ore"
-              },
-              "result": "minecraft:coal",
-              "experience": 0.1,
-              "cookingtime": 200
-            }
-            """;
+
+    private static final String EXPERIENCE = "experience";
+    private static final String COOKING_TIME = "cookingtime";
 
     @Test
     public void simpleCompareFirst() {
-        JsonObject first = TestUtils.json(recipe, j -> j.addProperty("experience", 0.2));
-        JsonObject second = TestUtils.json(recipe); // 0.1 experience
+        JsonObject first = TestUtils.json(TestUtils.Recipes.SMELTING, j -> j.addProperty(EXPERIENCE, 0.2));
+        JsonObject second = TestUtils.json(TestUtils.Recipes.SMELTING); // 0.1 experience
 
-        LinkedHashMap<String, JsonCompare.Rule> rules = new LinkedHashMap<>();
-        rules.put("experience", new JsonCompare.LowerRule());
+        Map<String, JsonCompare.Rule> rules = new LinkedHashMap<>();
+        rules.put(EXPERIENCE, new JsonCompare.LowerRule());
         JsonObject result = JsonCompare.compare(rules, first, second);
         assertEquals(second, result);
     }
 
     @Test
     public void simpleCompareSecond() {
-        JsonObject first = TestUtils.json(recipe, j -> j.addProperty("experience", 0.05));
-        JsonObject second = TestUtils.json(recipe); // 0.1 experience
+        JsonObject first = TestUtils.json(TestUtils.Recipes.SMELTING, j -> j.addProperty(EXPERIENCE, 0.05));
+        JsonObject second = TestUtils.json(TestUtils.Recipes.SMELTING); // 0.1 experience
 
-        LinkedHashMap<String, JsonCompare.Rule> rules = new LinkedHashMap<>();
-        rules.put("experience", new JsonCompare.LowerRule());
+        Map<String, JsonCompare.Rule> rules = new LinkedHashMap<>();
+        rules.put(EXPERIENCE, new JsonCompare.LowerRule());
         JsonObject result = JsonCompare.compare(rules, first, second);
         assertEquals(first, result);
     }
 
     @Test
     public void compareHigherWins() {
-        JsonObject first = TestUtils.json(recipe, j -> j.addProperty("experience", 0.05));
-        JsonObject second = TestUtils.json(recipe); // 0.1 experience  // 0.1 experience
+        JsonObject first = TestUtils.json(TestUtils.Recipes.SMELTING, j -> j.addProperty(EXPERIENCE, 0.05));
+        JsonObject second = TestUtils.json(TestUtils.Recipes.SMELTING); // 0.1 experience
 
-        LinkedHashMap<String, JsonCompare.Rule> rules = new LinkedHashMap<>();
-        rules.put("experience", new JsonCompare.HigherRule());
+        Map<String, JsonCompare.Rule> rules = new LinkedHashMap<>();
+        rules.put(EXPERIENCE, new JsonCompare.HigherRule());
         JsonObject result = JsonCompare.compare(rules, first, second);
         assertEquals(second, result);
     }
 
     @Test
     public void compareMulti() {
-        JsonObject a = TestUtils.json(recipe, j -> {
-            j.addProperty("experience", 0.1);
-            j.addProperty("cookingtime", 100);
+        JsonObject a = TestUtils.json(TestUtils.Recipes.SMELTING, j -> {
+            j.addProperty(EXPERIENCE, 0.1);
+            j.addProperty(COOKING_TIME, 100);
         });
-        JsonObject b = TestUtils.json(recipe, j -> j.addProperty("experience", 0.1));
-        JsonObject c = TestUtils.json(recipe, j -> {
-            j.addProperty("experience", 0.1);
-            j.addProperty("cookingtime", 50);
+        JsonObject b = TestUtils.json(TestUtils.Recipes.SMELTING, j -> j.addProperty(EXPERIENCE, 0.1));
+        JsonObject c = TestUtils.json(TestUtils.Recipes.SMELTING, j -> {
+            j.addProperty(EXPERIENCE, 0.1);
+            j.addProperty(COOKING_TIME, 50);
         });
-        JsonObject d = TestUtils.json(recipe, j -> j.addProperty("experience", 0.2));
-        JsonObject e = TestUtils.json(recipe, j -> j.addProperty("experience", 0.2));
-        JsonObject f = TestUtils.json(recipe, j -> j.addProperty("experience", 0.1));
-        JsonObject g = TestUtils.json(recipe, j -> {
-            j.addProperty("experience", 0.2);
-            j.addProperty("cookingtime", 100);
+        JsonObject d = TestUtils.json(TestUtils.Recipes.SMELTING, j -> j.addProperty(EXPERIENCE, 0.2));
+        JsonObject e = TestUtils.json(TestUtils.Recipes.SMELTING, j -> j.addProperty(EXPERIENCE, 0.2));
+        JsonObject f = TestUtils.json(TestUtils.Recipes.SMELTING, j -> j.addProperty(EXPERIENCE, 0.1));
+        JsonObject g = TestUtils.json(TestUtils.Recipes.SMELTING, j -> {
+            j.addProperty(EXPERIENCE, 0.2);
+            j.addProperty(COOKING_TIME, 100);
         });
 
-        LinkedHashMap<String, JsonCompare.Rule> rules = new LinkedHashMap<>();
-        rules.put("experience", new JsonCompare.HigherRule());
-        rules.put("cookingtime", new JsonCompare.LowerRule());
+        Map<String, JsonCompare.Rule> rules = new LinkedHashMap<>();
+        rules.put(EXPERIENCE, new JsonCompare.HigherRule());
+        rules.put(COOKING_TIME, new JsonCompare.LowerRule());
 
         List<JsonObject> list = Arrays.asList(a, b, c, d, e, f, g);
         list.sort((first, second) -> JsonCompare.compare(first, second, rules));
@@ -91,116 +83,63 @@ public class JsonCompareTest {
 
     @Test
     public void simpleMatch() {
-        JsonObject first = TestUtils.json(recipe);
-        JsonObject second = TestUtils.json(recipe);
-        boolean matches = JsonCompare.matches(first, second, List.of());
+        JsonObject first = TestUtils.json(TestUtils.Recipes.SMELTING);
+        JsonObject second = TestUtils.json(TestUtils.Recipes.SMELTING);
+        boolean matches = JsonCompare.matches(first, second, TestUtils.DEFAULT_COMPARE_SETTINGS);
         assertTrue(matches);
     }
 
     @Test
     public void noMatch() {
-        JsonObject first = TestUtils.json(recipe, j -> j.addProperty("experience", 100));
-        JsonObject second = TestUtils.json(recipe);
-        boolean matches = JsonCompare.matches(first, second, List.of());
+        JsonObject first = TestUtils.json(TestUtils.Recipes.SMELTING, j -> j.addProperty(EXPERIENCE, 100));
+        JsonObject second = TestUtils.json(TestUtils.Recipes.SMELTING);
+        boolean matches = JsonCompare.matches(first, second, new JsonCompare.CompareSettings());
         assertFalse(matches);
     }
 
     @Test
     public void matchBecauseIgnore() {
-        JsonObject first = TestUtils.json(recipe, j -> j.addProperty("experience", 100));
-        JsonObject second = TestUtils.json(recipe);
-        boolean matches = JsonCompare.matches(first, second, List.of("experience"));
+        JsonObject first = TestUtils.json(TestUtils.Recipes.SMELTING, j -> j.addProperty(EXPERIENCE, 100));
+        JsonObject second = TestUtils.json(TestUtils.Recipes.SMELTING);
+        var compareSettings = TestUtils.getDefaultCompareSettings();
+        compareSettings.ignoreField(EXPERIENCE);
+        boolean matches = JsonCompare.matches(first, second, compareSettings);
         assertTrue(matches);
     }
 
     @Test
     public void shapedNoMatch() {
-        String recipe1 = """
-                {
-                  "type": "minecraft:crafting_shaped",
-                  "pattern": [
-                    "iii",
-                    "ici",
-                    "iii"
-                  ],
-                  "key": {
-                    "i": {
-                      "tag": "forge:raw_materials/iron"
-                    },
-                    "k": {
-                      "item": "minecraft:carrot"
-                    }
-                  },
-                  "result": "minecraft:iron_ingot"
-                }
-                """;
-        String recipe2 = """
-                {
-                  "type": "minecraft:crafting_shaped",
-                  "pattern": [
-                    "iii",
-                    "ici",
-                    "iii"
-                  ],
-                  "key": {
-                    "i": {
-                      "tag": "forge:raw_materials/iron"
-                    },
-                    "k": {
-                      "item": "minecraft:pumpkin"
-                    }
-                  },
-                  "result": "minecraft:iron_ingot"
-                }
-                """;
-
-        JsonObject first = TestUtils.json(recipe1);
-        JsonObject second = TestUtils.json(recipe2);
-        JsonObject result = JsonCompare.compareShaped(first, second, List.of("pattern", "key"));
+        JsonObject first = TestUtils.json(TestUtils.Recipes.SHAPED_NO_MATCH_1);
+        JsonObject second = TestUtils.json(TestUtils.Recipes.SHAPED_NO_MATCH_2);
+        JsonObject result = JsonCompare.compareShaped(first, second, TestUtils.DEFAULT_SHAPED_COMPARE_SETTINGS);
         assertNull(result);
     }
 
     @Test
-    public void shapedMatch() {
-        String recipe1 = """
-                {
-                  "type": "minecraft:crafting_shaped",
-                  "pattern": [
-                    "iii",
-                    "iii",
-                    "iii"
-                  ],
-                  "key": {
-                    "i": {
-                      "tag": "forge:raw_materials/iron"
-                    }
-                  },
-                  "result": "minecraft:iron_ingot"
-                }
-                """;
-        String recipe2 = """
-                {
-                  "type": "minecraft:crafting_shaped",
-                  "pattern": [
-                    "iii",
-                    "iki",
-                    "iii"
-                  ],
-                  "key": {
-                    "i": {
-                      "tag": "forge:raw_materials/iron"
-                    },
-                    "k": {
-                      "tag": "forge:raw_materials/iron"
-                    }
-                  },
-                  "result": "minecraft:iron_ingot"
-                }
-                """;
-
-        JsonObject first = TestUtils.json(recipe1);
-        JsonObject second = TestUtils.json(recipe2);
-        JsonObject result = JsonCompare.compareShaped(first, second, List.of("pattern", "key"));
+    public void shapedSpecialMatch() {
+        JsonObject first = TestUtils.json(TestUtils.Recipes.SHAPED_SPECIAL_MATCH_1);
+        JsonObject second = TestUtils.json(TestUtils.Recipes.SHAPED_SPECIAL_MATCH_2);
+        JsonObject result = JsonCompare.compareShaped(first, second, TestUtils.DEFAULT_SHAPED_COMPARE_SETTINGS);
         assertEquals(first, result);
+    }
+
+    @Test
+    public void sanitizeImplicitCount() {
+        JsonObject first = TestUtils.json(TestUtils.Recipes.SHAPED_SANITIZE_1);
+        JsonObject second = TestUtils.json(TestUtils.Recipes.SHAPED_SANITIZE_2);
+        var compareSettings = TestUtils.getDefaultShapedCompareSettings();
+        compareSettings.setShouldSanitize(true);
+        JsonObject result = JsonCompare.compareShaped(first, second, compareSettings);
+        assertEquals(first, result);
+    }
+
+    @Test
+    public void sanitizeImplicitCountNested() {
+        JsonObject first = TestUtils.json(TestUtils.Recipes.CRUSHING_NESTED_SANITIZE_1);
+        JsonObject second = TestUtils.json(TestUtils.Recipes.CRUSHING_NESTED_SANITIZE_2);
+        var compareSettings = TestUtils.getDefaultCompareSettings();
+        compareSettings.setShouldSanitize(true);
+        boolean result = JsonCompare.matches(first, second, compareSettings);
+        assertTrue(result);
     }
 }